Add navbar skeleton, change main to header.scss, add sanduba logo
The sanduba logo can be later replaced by better assets. The current
attempt is to do the navbar in the way the exercise asks for. Later on
I wanna add bootstrap to this project but doing it manually should be
better as a way to practice RN.

The overall idea of the project now is to take the sanduba page as an
inspiration and hopefully build one that looks a bit better if possible