Add content skeleton to start iterating colors/designs
Added a mockup of the content of the landing page that'll load at the
start of opening our page. Should use it to start iterating with the
colors shapes and organization of the final page.

Changed the header also, put the logo at the beginning and added some
padding to it.